Added sqlite db, and upgraded dotnet 9.0.0 to 10.0.1.
