+++ Server/appsettings.json
@@ -8,6 +8,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "DefaultConnection": "Server=localhost;Database=PokemonTCG;Integrated Security=true;TrustServerCertificate=true;"
+    "DefaultConnection": "Data Source=Data\\PokemonDb.db"
   }
 }
